requirements.txt
docopt>=0.6.2
pandas>=0.16.0
python-Levenshtein>=0.12.0
numpy>=1.9.2